Add various default definitions for freestanding targets
When compiling druntime for freestanding targets - IE targets with an
unknown OS in the triple - some definitions are missing. This patch uses
some definitions in the Newlib implementation of libc (typically used
without an OS) to provide relatively sensible defaults.

Future work would be to add Newlib as a version identifier and rewrite
the driver to detect the actual libc version rather than using the OS in
the triple to determine the libc used.